👷 Simplified timepicker to satisfy build
